♻️ refactor: 기존 JS/JSX 파일을 TS/TSX로 변환 및 구성 파일 추가 (#6)
- custom.d.ts 파일 추가하여 이미지 모듈 선언
- 사용하지 않는 react.svg 파일 삭제
