+++ src/custom.d.ts
@@ -1,0 +1,9 @@
+declare module '*.jpg' {
+  const value: string;
+  export default value;
+}
+
+declare module '*.png' {
+  const value: string;
+  export default value;
+}
