+++ src/pages/GitPage.tsx
@@ -1,0 +1,14 @@
+function GitPage() {
+  return (
+    <div className="page-container">
+      <h2>깃 사용법 - MVP 개발</h2>
+      <ul>
+        <li>Git 설치 및 초기화</li>
+        <li>git init / add / commit / push</li>
+        <li>GitHub 연동 실습</li>
+      </ul>
+    </div>
+  );
+}
+
+export default GitPage;
