--- src/pages/ProjectsPage.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import '../styles/Page.css';
-
-function ProjectsPage() {
-    return (
-        <div className="page-container">
-            <h2>프로젝트 모집 - 추후 개발 🔒 </h2>
-            <p>같이 프로젝트를 할 팀원을 모집해보세요!</p>
-            <button className="basic-button">팀 만들기</button>
-        </div>
-    );
-}
-
-export default ProjectsPage;
